+++ maven-plugin/src/test/java/de/redsix/dmncheck/CheckerMainTest.java
@@ -5,8 +5,6 @@ import org.apache.maven.project.MavenProject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junitpioneer.jupiter.TempDirectory;
 
 import java.io.File;
 import java.nio.file.Paths;