+++ validators/src/main/java/de/redsix/dmncheck/feel/FeelExpression.java
@@ -14,6 +14,7 @@ public abstract class FeelExpression {
 
     public interface Cases<R> {
         R Empty();
+        R Null();
         R BooleanLiteral(Boolean aBoolean);
         R DateLiteral(LocalDateTime dateTime);
         R DoubleLiteral(Double aDouble);
@@ -35,6 +36,7 @@ public abstract class FeelExpression {
     public boolean containsVariable(final String name) {
         return FeelExpressions.caseOf(this)
                 .Empty_(false)
+                .Null_(false)
                 .BooleanLiteral_(false)
                 .DateLiteral_(false)
                 .DoubleLiteral_(false)
@@ -45,6 +47,34 @@ public abstract class FeelExpression {
                 .UnaryExpression((__, expression) -> expression.containsVariable(name))
                 .BinaryExpression((left, __, right) -> left.containsVariable(name) || right.containsVariable(name))
                 .DisjunctionExpression((head, tail) -> head.containsVariable(name) || tail.containsVariable(name));
+    }
+
+    public boolean isLiteral() {
+        return FeelExpressions.caseOf(this)
+                       .BooleanLiteral_(true)
+                       .DateLiteral_(true)
+                       .DoubleLiteral_(true)
+                       .IntegerLiteral_(true)
+                       .StringLiteral_(true)
+                       .VariableLiteral_(true)
+                       .otherwise_(false);
+
+    }
+
+    public boolean containsNot() {
+        return FeelExpressions.caseOf(this)
+                              .Empty_(false)
+                              .Null_(false)
+                              .BooleanLiteral_(false)
+                              .DateLiteral_(false)
+                              .DoubleLiteral_(false)
+                              .IntegerLiteral_(false)
+                              .StringLiteral_(false)
+                              .VariableLiteral_(false)
+                              .RangeExpression((__, lowerBound, upperBound, ___) -> lowerBound.containsNot() || upperBound.containsNot())
+                              .UnaryExpression((operator, expression) -> operator.equals(Operator.NOT))
+                              .BinaryExpression((left, __, right) -> left.containsNot() || right.containsNot())
+                              .DisjunctionExpression((head, tail) -> head.containsNot() || tail.containsNot());
     }
 
     @Override